Visual style (#19)
* Begin style revisions

* Update index.html.nj

* sort on _rank

* Update index.html.nj

* Further visual changes

Mostly for the smaller browser sizes.

* separate values and HTML when sorting the table

* add sorted on class

* Revise visual style and names of teams, add badges

All of the above. Not done – need to tweak the column widths for smaller screens.

* Accept changes, I think

* Nope, accept the incoming changes

Because “current” and “incoming” mean basically the same thing in this context. Seriously, who names this shit?

* Adjust the horizontal rules

* Adjust table cell spacing and alignment

* Adjust the clip-path for the logo-box

Still doesn’t seem to do anything in Safari. Also, Safari has its own weird way of interpreting the margins etc.

* Adjust table headings for smaller screen sizes